Move pkg/cloud/aws/client under pkg/client
The client is not actuatlly aws client, rather a wrapper built on top of aws services.
Thus, no need to mention cloud or aws in the package path.
