+++ cmd/aws-actuator/main.go
@@ -37,7 +37,7 @@ import (
 	"github.com/golang/glog"
 	"github.com/spf13/cobra"
 
-	awsclient "sigs.k8s.io/cluster-api-provider-aws/pkg/cloud/aws/client"
+	awsclient "sigs.k8s.io/cluster-api-provider-aws/pkg/client"
 	clusterv1 "sigs.k8s.io/cluster-api/pkg/apis/cluster/v1alpha1"
 
 	"github.com/ghodss/yaml"
